BT Starter

+++ src/org/andicar/persistence/MainDbAdapter.java
@@ -30,6 +30,8 @@ import java.util.ArrayList;
 import org.andicar.utils.StaticValues;
 import org.andicar.activity.R;
 import org.andicar.utils.AndiCarStatistics;
+
+import com.andicar.addon.persistence.AddOnDBObjectDef;
 public class MainDbAdapter extends DB
 {
     SharedPreferences mPref;
@@ -317,19 +319,12 @@ public class MainDbAdapter extends DB
                 isActiveContent.put(GEN_COL_ISACTIVE_NAME, content.getAsString(GEN_COL_ISACTIVE_NAME));
                 mDb.update(TASK_CAR_TABLE_NAME, isActiveContent, TASK_CAR_COL_CAR_ID_NAME + " = ?", whereArgs);
                 mDb.update(TODO_TABLE_NAME, isActiveContent, TODO_COL_CAR_ID_NAME + " = ?", whereArgs);
+                if(content.getAsString(GEN_COL_ISACTIVE_NAME).equals("N")){
+	                if(!AddOnDBObjectDef.recordUpdated(mDb, CAR_TABLE_NAME, rowId, content))
+	                	return R.string.ERR_063;
+                }
+                
             }
-//            else if(tableName.equals(TASK_TABLE_NAME)){ 
-//            	//inactivate/activate the related task-car links; delete todos (will be recreated based on new task definition)
-//                String[] whereArgs = {Long.toString(rowId)};
-//                ContentValues isActiveContent = new ContentValues();
-//                isActiveContent.put(GEN_COL_ISACTIVE_NAME, content.getAsString(GEN_COL_ISACTIVE_NAME));
-//                mDb.update(TASK_CAR_TABLE_NAME, isActiveContent, TASK_CAR_COL_TASK_ID_NAME + " = ?", whereArgs);
-//                mDb.delete(TODO_TABLE_NAME, TODO_COL_TASK_ID_NAME + " = ? AND " + TODO_COL_ISDONE_NAME + " = 'N'", whereArgs);
-//            }
-/*
-                    mDb.delete(TASK_CAR_TABLE_NAME, TASK_CAR_COL_CAR_ID_NAME + "=" + rowId, null);
-                    mDb.delete(TODO_TABLE_NAME, TODO_COL_CAR_ID_NAME + "=" + rowId, null);
- */
             if(mCarId != -1 && stopIndex != null){
                try{
                     mDb.beginTransaction();
@@ -716,6 +711,9 @@ public class MainDbAdapter extends DB
                     //also delete the locations
                     mDb.delete(TASK_CAR_TABLE_NAME, TASK_CAR_COL_CAR_ID_NAME + "=" + rowId, null);
                     mDb.delete(TODO_TABLE_NAME, TODO_COL_CAR_ID_NAME + "=" + rowId, null);
+                    if(!AddOnDBObjectDef.recordDeleted(mDb, CAR_TABLE_NAME, rowId))
+                    	return R.string.ERR_063;
+                    
                     checkVal = (-1 * mDb.delete(tableName, GEN_COL_ROWID_NAME + "=" + rowId, null ));
                 }
                 else if(tableName.equals(TASK_TABLE_NAME)){
